+++ src/main/java/com/winterfoodies/winterfoodies_project/dto/cart/CartDto.java
@@ -12,17 +12,17 @@ import java.util.ArrayList;
 @Setter
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class CartDto {
-    private Long id;
+    private double id;
 
     private String name;
 
     private String storeName;
 
-    private Long price;
+    private double price;
 
-    private Long quantity;
+    private double quantity;
 
-    private Long totalPrice;
+    private double totalPrice;
 
     private String EstimatedCookingTime;
