+++ src/main/java/com/winterfoodies/winterfoodies_project/dto/store/StoreRequestDto.java
@@ -43,7 +43,7 @@ public class StoreRequestDto {
 
     private double longitude;
 
-    private Long averageRating; // 평균 별점
+    private double averageRating; // 평균 별점
     
     private String estimatedCookingTime;
 }
